requirements.txt
camelot-py==0.9.0
PyPDF2==2.12.1
pandas==2.2.3
opencv-python==4.10.0.84
